change int to long in method to generateNumber

+++ src/main/java/com/library/controller/ApprovisionnementController.java
@@ -37,7 +37,7 @@ public class ApprovisionnementController {
     }
 
     @GetMapping("/searchApprovisionnementByCode")
-    public Approvisionnement getApprovisionnementByCode(@RequestParam("code") int code) {
+    public Approvisionnement getApprovisionnementByCode(@RequestParam("code") long code) {
         return approvisionnementService.findApprovisionnementByCode(code);
     }
